+++ udata/core/discussions/forms.py
@@ -1,7 +1,7 @@
 from udata.forms import ModelForm, Form, fields, validators
 from udata.i18n import lazy_gettext as _
 
-from .models import Discussion
+from .models import Discussion, COMMENT_SIZE_LIMIT
 
 __all__ = ('DiscussionCreateForm', 'DiscussionCommentForm')
 
@@ -11,12 +11,12 @@ class DiscussionCreateForm(ModelForm):
 
     title = fields.StringField(_('Title'), [validators.DataRequired()])
     comment = fields.StringField(
-        _('Comment'), [validators.DataRequired(), validators.Length(max=50000)])
+        _('Comment'), [validators.DataRequired(), validators.Length(max=COMMENT_SIZE_LIMIT)])
     subject = fields.ModelField(_('Subject'), [validators.DataRequired()])
     extras = fields.ExtrasField()
 
 
 class DiscussionCommentForm(Form):
     comment = fields.StringField(
-        _('Comment'), [validators.DataRequired(), validators.Length(max=50000)])
+        _('Comment'), [validators.DataRequired(), validators.Length(max=COMMENT_SIZE_LIMIT)])
     close = fields.BooleanField(default=False)